+++ dbdaora/sorted_set/_tests/conftest.py
@@ -1,21 +1,30 @@
+from dataclasses import dataclass
+
 import pytest
 
 from dbdaora import SortedSetEntity, SortedSetRepository
 
 
+@dataclass
+class FakeEntity(SortedSetEntity):
+    id: str
+
+
 class FakeRepository(SortedSetRepository[str]):
     entity_name = 'fake'
+    entity_id_name = 'id'
     key_attrs = ('id',)
+    entity_cls = FakeEntity
 
 
 @pytest.fixture
 def fake_entity():
-    return SortedSetEntity(id='fake', data=['1', '2'])
+    return FakeEntity(id='fake', values=['1', '2'])
 
 
 @pytest.fixture
 def fake_entity_withscores():
-    return SortedSetEntity(id='fake', data=[('1', 0), ('2', 1)])
+    return FakeEntity(id='fake', values=[('1', 0), ('2', 1)])
 
 
 @pytest.fixture