+++ dbdaora/sorted_set/entity.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from typing import Sequence, Union
+from typing import Any, Sequence, TypedDict, Union
 
 from dbdaora.data_sources.memory import RangeOutput
 
@@ -9,7 +9,18 @@ SortedSetInput = Sequence[Union[str, float]]
 SortedSetData = Union[RangeOutput, SortedSetInput]
 
 
-@dataclass
+@dataclass(init=False)
 class SortedSetEntity:
-    id: str
-    data: SortedSetData
+    values: SortedSetData
+
+    def __init__(self, values: SortedSetData, **kwargs: Any):
+        self.values = values
+
+
+class SortedSetDictEntity(TypedDict):
+    values: SortedSetData
+
+
+# see https://github.com/python/mypy/issues/4300
+# Entity = Union[SortedSetEntity, SortedSetDictEntity]
+Entity = Union[SortedSetEntity]